first step in merging easyfetch into easy2nfty
fetch_with_local is a test to combine the functionality of easyfetch with easy2ntfy for seamless transitions from intra to internet. it automatically detects when the app and the node are in the same local network and changes to direct requests to the node. most of it works already

todo:
- default local ip address as an input-field when the app is started in the local network. (right now it gets the ip from the gateway)
- the "dual commands" need to get split again for local network
- splitview functionality for ESPeasy webui in local mode