e2n_20230424/a7 fix: local detection
It seems to be impossible to make an reliable local network detection with "Mixed Content" being blocked since espeasys webserver doesn't do https.
The worst and only way right now to somehow detect if we are in the same network is to determine whenever a request through fetch is refused or timed out.
Therefore the request points to an server without https capabilities (espeasy) and gets refused...